crs_api/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM
openai==1.6.1
anthropic==0.7.0  # Optional: for Claude support
tiktoken==0.5.2

# Vector Store & ML
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Cache & Queue
redis==5.0.1
aioredis==2.0.1
celery==5.3.4

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Async & Performance
aiohttp==3.9.1
asyncio==3.4.3
tenacity==8.2.3
backoff==2.2.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
jsonschema==4.20.0
